# login_api/swagger.yml
swagger: "2.0"
info:
  title: API de Login
  description: Documentação da API de Autenticação e Usuários
  version: "1.0"
host: "3.238.233.30:5002"
basePath: "/"
schemes:
  - http
paths:
  /ping:
    get:
      tags:
        - Utilidade
      summary: Testa se a API está ativa
      responses:
        200:
          description: Retorna status OK
          schema:
            type: object
            properties:
              status:
                type: string
                example: OK

  /usuarios:
    get:
      tags:
        - Usuários
      summary: Lista todos os usuários cadastrados
      security:
        - Bearer: []
      responses:
        200:
          description: Lista de usuários
          schema:
            type: array
            items:
              $ref: "#/definitions/Usuario"
        401:
          description: Token não fornecido ou inválido
          schema:
            type: object
            properties:
              erro:
                type: string
                example: Token não fornecido

  /register:
    post:
      tags:
        - Autenticação
      summary: Registra um novo usuário
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UsuarioRegistro"
      responses:
        201:
          description: Usuário registrado com sucesso
          schema:
            type: object
            properties:
              mensagem:
                type: string
                example: Usuário registrado com sucesso!
              id:
                type: integer
                example: 1
              token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        400:
          description: Dados inválidos
          schema:
            type: object
            properties:
              erro:
                type: string
                example: Usuário e senha são obrigatórios
        409:
          description: Usuário já existe
          schema:
            type: object
            properties:
              erro:
                type: string
                example: Este nome de usuário já existe

  /login:
    post:
      tags:
        - Autenticação
      summary: Autentica um usuário
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UsuarioLogin"
      responses:
        200:
          description: Login bem-sucedido
          schema:
            type: object
            properties:
              mensagem:
                type: string
                example: Login bem-sucedido!
              user_id:
                type: integer
                example: 1
              username:
                type: string
                example: usuario123
              token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        400:
          description: Dados inválidos
          schema:
            type: object
            properties:
              erro:
                type: string
                example: Usuário e senha são obrigatórios
        401:
          description: Credenciais inválidas
          schema:
            type: object
            properties:
              erro:
                type: string
                example: Credenciais inválidas

  /perfil:
    get:
      tags:
        - Usuários
      summary: Retorna informações do usuário autenticado
      security:
        - Bearer: []
      responses:
        200:
          description: Informações do perfil do usuário
          schema:
            type: object
            properties:
              user_id:
                type: integer
                example: 1
              username:
                type: string
                example: usuario123
              mensagem:
                type: string
                example: Você está autenticado!
        401:
          description: Token não fornecido, inválido ou expirado
          schema:
            type: object
            properties:
              erro:
                type: string
                example: Token não fornecido

  /verificar-token:
    get:
      tags:
        - Autenticação
      summary: Verifica se o token JWT é válido
      security:
        - Bearer: []
      responses:
        200:
          description: Token válido
          schema:
            type: object
            properties:
              valido:
                type: boolean
                example: true
              user_id:
                type: integer
                example: 1
              username:
                type: string
                example: usuario123
        401:
          description: Token não fornecido, inválido ou expirado
          schema:
            type: object
            properties:
              erro:
                type: string
                example: Token não fornecido

definitions:
  Usuario:
    type: object
    properties:
      id:
        type: integer
        example: 1
      username:
        type: string
        example: usuario123

  UsuarioRegistro:
    type: object
    required:
      - username
      - password
    properties:
      username:
        type: string
        example: usuario123
      password:
        type: string
        format: password
        example: senhaSegura123

  UsuarioLogin:
    type: object
    required:
      - username
      - password
    properties:
      username:
        type: string
        example: usuario123
      password:
        type: string
        format: password
        example: senhaSegura123

securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: Token JWT no formato "Bearer {token}"